+++ include/random.h
@@ -8,7 +8,7 @@ namespace Cascade {
 
 void random_seed(uint64_t seed);
 uint64_t random_uint64();
-size_t random_bit_nr(size_t start_bit_nr, size_t end_bit_nr);
+int random_bit_nr(int start_bit_nr, int end_bit_nr);
 
 } /* namespace Cascade */
