+++ src/algorithm.cpp
@@ -11,10 +11,10 @@ static Algorithms algorithms;
 static double min_estimated_bit_error_rate = 0.00001;
 
 Algorithm::Algorithm(std::string name, 
-                     unsigned nr_cascade_iterations,
-                     size_t (*block_size_function)(unsigned iteration_nr,
+                     int nr_cascade_iterations,
+                     int (*block_size_function)(int iteration_nr,
                                                    double estimated_bit_error_rate),
-                     unsigned nr_biconf_iterations,
+                     int nr_biconf_iterations,
                      bool biconf_error_free_streak,
                      bool biconf_correct_complement,
                      bool biconf_cascade,
@@ -43,7 +43,7 @@ Algorithm* Algorithm::get_by_name(std::string name)
     }
 }
 
-static size_t original_block_size_function(unsigned iteration_nr, double estimated_bit_error_rate)
+static int original_block_size_function(int iteration_nr, double estimated_bit_error_rate)
 {
     if (estimated_bit_error_rate < min_estimated_bit_error_rate) {
         estimated_bit_error_rate = min_estimated_bit_error_rate;
