+++ src/random.cpp
@@ -17,9 +17,9 @@ uint64_t random_uint64()
     return dist(mt);
 }
 
-size_t random_bit_nr(size_t start_bit_nr, size_t end_bit_nr)
+int random_bit_nr(int start_bit_nr, int end_bit_nr)
 {
-    static std::uniform_int_distribution<size_t> dist(start_bit_nr, end_bit_nr);
+    static std::uniform_int_distribution<int> dist(start_bit_nr, end_bit_nr);
     return dist(mt);
 }
