+++ tests/test_reconciliation.cpp
@@ -17,7 +17,7 @@ TEST (Reconciliation, test_1)
     double bit_error_rate = 0.1;
     noisy_key.apply_noise(bit_error_rate);
     // TODO: Better way of announcing bit errors
-    for (size_t bit_nr = 0; bit_nr < 16; ++bit_nr) {
+    for (int bit_nr = 0; bit_nr < 16; ++bit_nr) {
         if (correct_key.get_bit(bit_nr) != noisy_key.get_bit(bit_nr)) {
             DEBUG("Bit error in " << bit_nr);
         }
